requirements.txt
fastapi==0.70.0
httpx==0.21.1
python-jose==3.3.0
python-dotenv==0.19.1
pydantic==1.8.2
boto3==1.18.48
botocore==1.21.48
uvicorn==0.15.0
pytest-cov
gradio